Create cli.js and use it for report to the console
